+++ CHANGELOG.md
@@ -2,6 +2,72 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.0.0](https://github.com/dansmaculotte/laravel-newsletter/compare/v2.0.0...v3.0.0) (2020-11-30)
+
+
+### ⚠ BREAKING CHANGES
+
+* drop support of Illuminate/Support 6
+
+* update dependencies for Laravel 8 support ([624341d](https://github.com/dansmaculotte/laravel-newsletter/commit/624341d3d9cdb4615725e3b713dbb046bb198dec))
+
+## [3.0.0](https://github.com/dansmaculotte/laravel-newsletter/compare/v2.0.0...v3.0.0) (2020-11-30)
+
+
+### ⚠ BREAKING CHANGES
+
+* drop support of Illuminate/Support 6
+
+* update dependencies for Laravel 8 support ([624341d](https://github.com/dansmaculotte/laravel-newsletter/commit/624341d3d9cdb4615725e3b713dbb046bb198dec))
+
+## [3.0.0](https://github.com/dansmaculotte/laravel-newsletter/compare/v2.0.0...v3.0.0) (2020-11-30)
+
+
+### ⚠ BREAKING CHANGES
+
+* drop support of Illuminate/Support 6
+
+* update dependencies for Laravel 8 support ([3bec69a](https://github.com/dansmaculotte/laravel-newsletter/commit/3bec69a8fd0196be2dd981a788505c11e1aaa1cb))
+
+## [2.0.0](https://github.com/dansmaculotte/laravel-newsletter/compare/v1.2.1...v2.0.0) (2020-05-06)
+
+
+### ⚠ BREAKING CHANGES
+
+* drop support of laravel 5
+
+### Features
+
+* allow laravel 7 ([773f7e0](https://github.com/dansmaculotte/laravel-newsletter/commit/773f7e0))
+
+### [1.2.1](https://github.com/dansmaculotte/laravel-newsletter/compare/v1.2.0...v1.2.1) (2020-04-01)
+
+
+### Bug Fixes
+
+* add getLastError phpDoc method ([e378989](https://github.com/dansmaculotte/laravel-newsletter/commit/e37898986bdbc113d04143b8d6aafede9855fd78))
+
+## [1.2.0](https://github.com/dansmaculotte/laravel-newsletter/compare/v1.1.2...v1.2.0) (2020-04-01)
+
+
+### Features
+
+* add getLastError method ([ab826cd](https://github.com/dansmaculotte/laravel-newsletter/commit/ab826cde3a324b2a214e59e435f7d25f91462a7a))
+
+### [1.1.2](https://github.com/dansmaculotte/laravel-newsletter/compare/v1.1.1...v1.1.2) (2020-03-31)
+
+
+### Bug Fixes
+
+* **mailjet:** cast listId to int when compared in isSubscribed method ([459024c](https://github.com/dansmaculotte/laravel-newsletter/commit/459024c1783f6a8a6661c8ba0997d41d1016ff68))
+
+### [1.1.1](https://github.com/dansmaculotte/laravel-newsletter/compare/v1.1.0...v1.1.1) (2020-03-30)
+
+
+### Bug Fixes
+
+* increase connection timeout to avoid cUrl errors ([#3](https://github.com/dansmaculotte/laravel-newsletter/issues/3)) ([ccf6168](https://github.com/dansmaculotte/laravel-newsletter/commit/ccf61684117367de8a2777359de5544d825b4569))
+
 ## [1.1.0](https://github.com/dansmaculotte/laravel-newsletter/compare/v1.0.1...v1.1.0) (2020-03-24)